+++ thebleep/shells/bash.py
@@ -24,12 +24,13 @@ class Bash(Generic):
         # the shell it ran in.
         return '''
             function {name} () {{
+                TB_EXIT=$?;
                 TB_SHELL_ALIASES=$(alias);
                 TB_HISTORY=$(fc -ln -10);
                 {fit_transport}
                 TB_CAN_EDIT=; [ "${{BASH_VERSINFO[0]:-0}}" -ge 4 ] && TB_CAN_EDIT=1;
                 TB_CMD=$(
-                    TB_SHELL=bash TB_ALIAS={name} TB_CAN_EDIT="$TB_CAN_EDIT" TB_SHELL_ALIASES="$TB_SHELL_ALIASES" TB_HISTORY="$TB_HISTORY" {command} {argument_placeholder} "$@"
+                    TB_SHELL=bash TB_ALIAS={name} TB_EXIT="$TB_EXIT" TB_CAN_EDIT="$TB_CAN_EDIT" TB_SHELL_ALIASES="$TB_SHELL_ALIASES" TB_HISTORY="$TB_HISTORY" {command} {argument_placeholder} "$@"
                 );
                 TB_STATUS=$?;
                 if [ "$TB_STATUS" -eq {exit_edit} ]; then
@@ -38,7 +39,7 @@ class Bash(Generic):
                     eval "$TB_CMD";
                     {alter_history}
                 fi;
-                unset TB_SHELL_ALIASES TB_HISTORY TB_CMD TB_STATUS TB_CAN_EDIT;
+                unset TB_SHELL_ALIASES TB_HISTORY TB_CMD TB_STATUS TB_CAN_EDIT TB_EXIT;
             }}
         '''.format(
             name=alias_name,